split "input.*"
split-off "input_context.h" and "input_enums.h" from "input.h" and "input_context.cpp" from "input.cpp"
"input_enums.h" contains things that both of the other header files need.
Therefor it is included in both of the header files.
